Fix missing project dependency detection (#421)
* Add multiple target frameworks to ProjectA and ProjectB

This demonstrates that project dependency detection has a bug and makes ShouldDetectProjectChanges fail.

* Fix missing project dependency detection

When multiple target frameworks are defined, the dictionary of source file to project must keep all the project ids. Choosing just one with `.First()` can lead to failure because the project id might correspond to a target framework that does not match the target framework of the referencing project.

---------

Co-authored-by: Aaron Stannard <aaron@petabridge.com>